Added fill_in_contribution_to_jacobian_and_mass_matrix() to the PVDEquationsWithPressure so that bifurcations can be detected and eigenproblems can be solved for solid mechanics elements.
Reformulated the calculation of the Jacobian of solid elements so that the 
remaining finite-differencing of the constitutive law is handled by the
constitutive law and can be overloaded so that a completely analytic Jacobain
can be calculated. The default implementation is still finite differences.
Also added some slight optimisations so that the calculation is a little bit
faster and uses a bit less storage.