Further through the parallelisation of bifurcation-tracking routines
The pitchfork detection works in parallel in
non-block mode for distributed problems and in block mode for non-distributed
problems. The parallelisation has necessitated the introduction of a new
class DoubleVectorWithHalo which are typically "problem level" vectors 
for which information on other processors is required to construct the
augmented system used to detected the pitchfork. The setup of these structures
required the introduction of a few other classes in the Problem, Mesh, 
Elements and Nodes to return the global equation numbers of all possible
halo data.

A synchronise function has also been added to the AssemblyHandler class
and is called whenever synchronisation takes place within a newton step.
It is used to synchronise "global data" such as the bifurcation parameter. 
Some small rewrites of bits of the problem have also been performed to use
functions to calculate information that is required in a couple of different
contexts.