Modified jeffery_orbit test to use a global measure of the solution so that it should now pass even if the mesh varies between different processors.
Included Patrick's revision to handle analytic calculation of shape
derivatives ... this may still require some refactorisation/cleaning.

Modified the triangle mesh so that it can now dump and restart and added test
for this in navier_stokes/unstructured_fluid. This is not as tidy as I would like, but this revision must be checked in quickly because

 IMPORTANT:
 
 Added changes to static const member data to avoid a compiler bug(?) in 
 gcc 4.4.4. Such member data of templated classes appears not to be initialised
 correctly. It will be assumed to be zero, which is an error. I have avoided 
 majority of the problems by removing the static data where possible and using
 explicit instantiation elsewhere:

  template<unsigned DIM>
  const double MyClass<DIM>::Whatever = 1.0;

  is replaced by

  template<>
  const double MyClass<1>::Whatever = 1.0;
  template<>
  const double MyClass<2>::Whatever = 1.0;
  template<>
  const double MyClass<3>::Whatever = 1.0;