Check in after trip to Sandia:
1) Added interface for (parallel and serial) eigensolves using the Trilinos
 Anasazi package. Tested in mpi/distribution/eigenproblem (parallel) 
 and eigenproblem (serial). This involved creating a new DoubleMultiVector
 class which is not yet as efficient as it should be.

2) Fixed a funny bug in LinearAlgebraDistribution which only showed up
on my machine. A particular construtor labelled the data as distributed even
if there was only one processor on the communicator, which messed things up.
What worries me is that the tests passed on the wulfling, so I fear that there
may be another more subtle bug lurking...

3) Added basic periodic orbit tracking capability, tested in 
bifurcation_tracking/periodic_orbit.cc. Again not as efficient as the ultimate
implementation should be, but is a working proof of concept. Need to work
on better temporal error estimators.

4) Fixed an oversight in frontal_solver.cc that allocated memory from stack
instead of heap and so died when elements had large numbers of dofs.

5) Relabelled a compiler macro in tetgen to avoid a clash with Anasazi.

6) Sightly hacked the immersed rigid bodies to better solve Tom's problem.