Fixed a bug in the check_links_in_example_list script so that it can now be run in any directory.
Added instructions about compiling Trilinos version 10 (and higher) using the
new build machinery.

Modified the spherical navier stokes equations so that they can use the
latest version of the LSC preconditioners.

Updated the mesh_as_geometric_object.cc to issue a warning if the number
of elements per bin is too large (greater than 100) and increase the number
of bins in the demo unstructured_adaptive_fsi.cc so that it runs at a  
sensible speed again.