Included proof-of-concept code for spatially adaptive continuation, but the method will not work for problems in which the Nodes move.
A complete rewrite of the scheme is planned for the remainder of this week,
in which the additional data is stored at the Data rather than the Problem 
level. This will allow seamless use of the spatial adaptivity with minimum
overhead.

In addition, corrected a few bugs in the dump and reload, which meant that
it didn't work if the min_permitted error was set to exactly 0.0