Small modification to fix an unintended bug :
 If Nodes were shared between submeshes the equation numbering could have
gone wrong. We never do/did this, so the problem was never spotted. 

 The problem has been fixed by assigning global equation numbers on the 
global mesh (only) and explicitly treating spines differently 
(which are something of an exception to most of our rules anyway). 
In addition the virtual status of 
Problem::assign_eqn_numbers() and Mesh::assign_local_eqn_numbers() have been 
removed because they are not needed.