block_preconditioner.h - re-implemented Richard's binary seach function in terms of STL functions.
I also renamed it to get_index_of_value instead of get_index_of_element.

There is an optional parameter for the user to indicate if the vector is
sorted or not. The old function only works if the vector is sorted, this
is not the case anymore! Both version is tested and it works.