Remove regenerate_config hacks from build machinery
Now running autogen.sh should automatically select the required build
steps in every case! Everything should just work and you should never
need to run it with the --rebuild option.

This will also massively speed up rerunning autogen in many cases
because ./configure is only run if it is really needed (previously it
was always run).

We are now almost using autotools in the "approved" way,
except that we generate the list of user directories automatically and
allow specification of configure options from a file.

The changes needed were:

* Include the directory list file into configure.ac properly.

* Use `autoreconf -i` instead of attempting (and failing) to run the
  right selection of autotools commands manually. This allows make to
  automatically detect when the driver list file has changed and rerun
  the necessary commands.

* Add code to autogen.sh to detect when the configure options have
  changed and force the rerunning of autoreconf automatically (instead
  of needing to specify a rebuild).

* Remove regenerate_config_files.sh and related files because they are
  no longer needed.