Fixed validate scripts for two_d_multi_poisson.
There were lots of dummy oks missing.
In mpi/solvers/ we use fpdiff.py for the convergence history. This is
because the driver code outputs the number of iterations as a whole
number. So testing the number of lines in the file does not make sense.

I have unzipped the following file:
new file:   linear_solvers/validata/one_plus_four_with_two_coarse_iterative_solver_convergence.dat
deleted:    linear_solvers/validata/one_plus_four_with_two_coarse_iterative_solver_convergence.dat.gz

because the line comparison script is not smart enough to unzip it for
me, this was the cause of the failed test.